[bug-fix] shmat return test on >(MAX_INT) machines
git-svn-id: http://svn.code.sf.net/p/rts-2/code/trunk/rts-2@1267 65ab8f4e-f147-0410-b3a9-f14133ecfe55

+++ src/utils/devser.c
@@ -1019,7 +1019,7 @@ void *
 devser_shm_data_at ()
 {
   void *ret = shmat (data_shm, NULL, 0);
-  if ((int) ret < 0)
+  if (ret < 0)
     {
       syslog (LOG_ERR, "devser_shm_data_at shmat: %m");
       exit (EXIT_FAILURE);
